move tasks, tests, fixtures, to sites app
