+++ server/apps/sites/services/__init__.py
@@ -1,2 +1,3 @@
 from .epa_site_service import EpaSiteService
 from .profile_service import EpaProfileService
+from .site_service import SiteService
