+++ server/apps/sites/tasks/__init__.py
@@ -1,0 +1,3 @@
+from .epa_site_task import get_epa_site
+from .profile_task import sync_user_sites
+from .site_task import sync_site_manifests
