+++ server/apps/sites/tasks/site_task.py
@@ -9,7 +9,7 @@ logger = logging.getLogger(__name__)
 @shared_task(name="sync site manifests", bind=True)
 def sync_site_manifests(self, *, site_id: str, username: str):
     """asynchronous task to sync an EPA site's manifests"""
-    from apps.trak.services import SiteService
+    from apps.sites.services import SiteService
 
     try:
         site_service = SiteService(username=username)
