+++ server/apps/sites/tests/test_serializers.py
@@ -3,18 +3,31 @@ import json
 import pytest
 
 from apps.sites.models import SitePermission
-from apps.sites.serializers import SitePermissionSerializer
+from apps.sites.serializers import (
+    ContactSerializer,
+    EpaPermissionSerializer,
+    EpaSiteSerializer,
+    SitePermissionSerializer,
+)
 
 
 class TestContactSerializer:
+    @pytest.fixture
+    def contact_serializer(self, haztrak_json) -> ContactSerializer:
+        return ContactSerializer(data=haztrak_json.CONTACT.value)
+
     def test_serializes(self, contact_serializer) -> None:
         contact_serializer.is_valid()
         assert contact_serializer.is_valid() is True
 
 
 class TestEpaSiteSerializer:
-    def test_serializes(self, handler_serializer):
-        assert handler_serializer.is_valid() is True
+    @pytest.fixture
+    def epa_site_serializer(self, haztrak_json) -> EpaSiteSerializer:
+        return EpaSiteSerializer(data=haztrak_json.HANDLER.value)
+
+    def test_serializes(self, epa_site_serializer):
+        assert epa_site_serializer.is_valid() is True
 
 
 class TestSitePermissionSerializer:
@@ -26,6 +39,10 @@ class TestSitePermissionSerializer:
     @pytest.fixture(autouse=True)
     def _permissions(self, site_permission_factory):
         self.permissions = site_permission_factory()
+
+    @pytest.fixture
+    def site_permission_serializer(self, haztrak_json) -> SitePermissionSerializer:
+        return SitePermissionSerializer(data=haztrak_json.SITE_PERMISSION.value)
 
     def test_serializes(self, site_permission_serializer) -> None:
         assert site_permission_serializer.is_valid() is True
@@ -45,8 +62,12 @@ class TestEpaPermissionSerializer:
     currently we don't serialize, only deserialize
     """
 
+    @pytest.fixture
+    def epa_permission_serializer(self, haztrak_json) -> EpaPermissionSerializer:
+        return EpaPermissionSerializer(data=haztrak_json.EPA_PERMISSION.value)
+
     def test_deserializes_epa_permissions(
-        self, epa_permission_serializer, rcra_profile_factory, site_factory
+        self, epa_permission_serializer, epa_profile_factory, site_factory
     ) -> None:
         if not epa_permission_serializer.is_valid():
             # if something is wrong with the serializer fixture, fail
@@ -54,5 +75,5 @@ class TestEpaPermissionSerializer:
         SitePermission.objects.create(
             **epa_permission_serializer.validated_data,
             site=site_factory(),
-            profile=rcra_profile_factory(),
+            profile=epa_profile_factory(),
         )
