+++ server/apps/trak/tasks/__init__.py
@@ -1,5 +1,2 @@
-from .handler_task import get_epa_site
 from .lookup_task import pull_federal_codes
 from .manifest_task import pull_manifest, sign_manifest
-from .profile_task import sync_user_sites
-from .site_task import sync_site_manifests
