+++ server/apps/trak/tests/models/test_handler.py
@@ -1,32 +1,29 @@
 import pytest
 
-from apps.sites.models import EpaSite
 from apps.trak.models import ManifestHandler
 
 
 @pytest.mark.django_db
 class TestManifestHandlerModel:
-    """Test related to the Manifest EpaSite model and its API"""
+    """Test related to the Manifest Handler model and its API"""
 
-    def test_handler_created(self, manifest_handler_factory) -> None:
+    def test_manifest_handler_created(self, manifest_handler_factory) -> None:
         assert type(manifest_handler_factory()) is ManifestHandler
 
-    def test_manager_creates_new_handler(self, handler_serializer) -> None:
-        handler_serializer.is_valid()
-        ManifestHandler.objects.save(epa_site=handler_serializer.validated_data)
-        new_handler = EpaSite.objects.get(epa_id=handler_serializer.validated_data["epa_id"])
-        assert isinstance(new_handler, EpaSite)
-
-    def test_signed_both_signatures_exists(self, manifest_handler_factory, e_signature_factory):
-        manifest_handler = manifest_handler_factory()
+    def test_is_signed_with_both_signatures_types(
+        self, manifest_handler_factory, e_signature_factory, paper_signature_factory
+    ):
+        manifest_handler = manifest_handler_factory(paper_signature=paper_signature_factory())
         e_signature_factory(manifest_handler=manifest_handler)
         assert manifest_handler.signed is True
 
-    def test_signed_only_paper(self, manifest_handler_factory):
+    def test_is_signed_when_paper_signature_exists(self, manifest_handler_factory):
         manifest_handler = manifest_handler_factory()
         assert manifest_handler.signed is True
 
-    def test_signed_only_electronic(self, epa_site_factory, e_signature_factory):
+    def test_is_signed_when_electronic_signature_exists(
+        self, epa_site_factory, e_signature_factory
+    ):
         manifest_handler = ManifestHandler(
             epa_site=epa_site_factory(),
             paper_signature=None,
@@ -35,7 +32,7 @@ class TestManifestHandlerModel:
         e_signature_factory(manifest_handler=manifest_handler)
         assert manifest_handler.signed is True
 
-    def test_signed_with_no_signatures(self, epa_site_factory):
+    def test_is_not_signed_with_no_signatures(self, epa_site_factory):
         manifest_handler = ManifestHandler(
             epa_site=epa_site_factory(),
             paper_signature=None,
