+++ server/apps/trak/tests/services/test_manifest_services.py
@@ -1,3 +1,4 @@
+from http import HTTPStatus
 from typing import Dict, List
 
 import pytest
@@ -17,6 +18,27 @@ class TestManifestService:
         self.gen001 = site_factory()
         self.json_100031134elc = haztrak_json.MANIFEST.value
         self.tracking_number = self.json_100031134elc.get("manifestTrackingNumber", "123456789ELC")
+
+    @pytest.fixture
+    def manifest_100033134elc_rcra_response(self, haztrak_json, mock_responses):
+        rcrainfo = RcrainfoService(api_username="testuser1", rcrainfo_env="preprod")
+        manifest_json = haztrak_json.MANIFEST.value
+        mock_responses.get(
+            url=f'{rcrainfo.base_url}/api/v1/emanifest/manifest/{manifest_json.get("manifestTrackingNumber")}',
+            content_type="application/json",
+            json=manifest_json,
+            status=HTTPStatus.OK,
+        )
+
+    @pytest.fixture
+    def search_site_mtn_rcra_response(self, haztrak_json, mock_responses):
+        rcrainfo = RcrainfoService(api_username="testuser1", rcrainfo_env="preprod")
+        mock_responses.post(
+            url=f"{rcrainfo.base_url}/api/v1/emanifest/search",
+            content_type="application/json",
+            json=[haztrak_json.MANIFEST.value.get("manifestTrackingNumber")],
+            status=HTTPStatus.OK,
+        )
 
     def test_pull_manifests(
         self, manifest_100033134elc_rcra_response, mocker: pytest_mock.MockerFixture
